[refactor]: rename providers_symphonie module class file names snake_case
